Added Table to Each Room
using java swing library JTable. Need to sync with a database possibly
in order to get new chores from admin. Admin adds chores and also checks
off chores whenever they are finished. If database is to expensive or
not possible then pull in from a file that is on the web or from a
dropbox file that way if the devices are synced then the app is up to
date.